requirements.txt
crypto==1.4.1
cryptography==37.0.2
fastapi==0.97.0
fastapi-jwt-auth==0.5.0
flasgger==0.9.5
Flask==2.1.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
httptools==0.5.0
idna==3.3
itsdangerous==2.1.2
janus==1.0.0
Jinja2==3.1.2
jsonschema==4.6.0
mypy-extensions==0.4.3
mysql-connector-python==8.0.29
mysqlclient==2.1.1
numpy==1.23.1
packaging==21.3
paho-mqtt==1.6.1
pandas==1.4.3
pathspec==0.10.3
platformdirs==2.5.2
prompt-toolkit==3.0.31
protobuf==3.20.1
psutil==5.9.4
pycparser==2.21
pycrypto==2.6.1
pycryptodome==3.14.1
pycryptodomex==3.14.1
pydantic==1.10.9
PyJWT==1.7.1
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyotp==2.8.0
pyparsing==3.0.9
pypng==0.20220715.0
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-http-client==3.3.7
pytz==2022.1
PyYAML==6.0
pyzmq==25.0.2
requests==2.27.1
requests-unixsocket==0.3.0
roundrobin==0.0.4
schedule==1.1.0
sendgrid==6.9.7
shellescape==3.8.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.16
starkbank-ecdsa==2.0.3
starlette==0.27.0
swagger-ui-bundle==0.0.9
typing_extensions==4.3.0
urllib3==1.26.9
uvicorn==0.22.0
uvloop==0.17.0
uWSGI==2.0.20
virtualenv==20.14.1
watchfiles==0.19.0
wcwidth==0.2.5
websockets==11.0.3
Werkzeug==2.1.2
